+++ functions/linebuilding.go
@@ -4,13 +4,14 @@ import (
 	"github.com/blutspende/go-astm/v2/constants/astmconst"
 	"github.com/blutspende/go-astm/v2/errmsg"
 	"github.com/blutspende/go-astm/v2/models"
+	"github.com/blutspende/go-astm/v2/models/astmmodels"
 	"math"
 	"reflect"
 	"strconv"
 	"time"
 )
 
-func BuildLine(sourceStruct interface{}, lineTypeName string, sequenceNumber int, config *models.Configuration) (result string, err error) {
+func BuildLine(sourceStruct interface{}, lineTypeName string, sequenceNumber int, config *astmmodels.Configuration) (result string, err error) {
 	// Process the target structure
 	sourceTypes, sourceValues, sourceTypesLength, err := ProcessStructReflection(sourceStruct)
 	if err != nil {
@@ -124,7 +125,7 @@ func BuildLine(sourceStruct interface{}, lineTypeName string, sequenceNumber int
 	return result, nil
 }
 
-func buildSubstructure(sourceStruct interface{}, config *models.Configuration) (result string, err error) {
+func buildSubstructure(sourceStruct interface{}, config *astmmodels.Configuration) (result string, err error) {
 	// Process the target structure
 	sourceTypes, sourceValues, sourceTypesLength, err := ProcessStructReflection(sourceStruct)
 	if err != nil {
@@ -183,7 +184,7 @@ func constructResult(fieldMap map[int]string, delimiter string, notation string)
 	return result
 }
 
-func convertField(field reflect.Value, annotation models.AstmFieldAnnotation, config *models.Configuration) (result string, err error) {
+func convertField(field reflect.Value, annotation models.AstmFieldAnnotation, config *astmmodels.Configuration) (result string, err error) {
 	// Check if the field is a pointer, nil returns empty, otherwise dereference it
 	if field.Kind() == reflect.Ptr {
 		if field.IsNil() {