+++ models/parsing.go
@@ -1,7 +1,5 @@
 package models
 
-import "time"
-
 // Annotation types for ASTM fields and structures
 type AstmFieldAnnotation struct {
 	Raw               string
@@ -22,25 +20,4 @@ type AstmStructAnnotation struct {
 	IsArray      bool
 	HasAttribute bool
 	Attribute    string
-}
-
-// Delimiters used in ASTM parsing
-type Delimiters struct {
-	Field     string
-	Repeat    string
-	Component string
-	Escape    string
-}
-
-var DefaultDelimiters = Delimiters{
-	Field:     `|`,
-	Repeat:    `\`,
-	Component: `^`,
-	Escape:    `&`,
-}
-
-// Internal configuration for parsing (part of the Configuration struct)
-type InternalConfiguration struct {
-	TimeLocation *time.Location
-	Delimiters   Delimiters
 }