+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.0] - 2025-04-25
+
+### Added
+- Record substructure support
+- Pointer field value support
+- Configuration parameter
+- Unit tests for modular functions
+- All charmap encoding support
+
+### Changed
+- Modular redesign
+- Message format structures, namings, annotations
+- Component array handling
+- Package names
+- Readme details
+
+### Fixed
+- Misannotated component parsing
+- Message format array singular names
+
 ## [0.9.4] - 2022-06-27
 
 ### Fixed